+++ packages/dataviews/src/components/dataform-controls/index.tsx
@@ -22,6 +22,7 @@ import select from './select';
 import text from './text';
 import toggle from './toggle';
 import textarea from './textarea';
+import richtext from './richtext';
 import toggleGroup from './toggle-group';
 import array from './array';
 import color from './color';
@@ -51,6 +52,7 @@ const FORM_CONTROLS: FormControls = {
 	text,
 	toggle,
 	textarea,
+	richtext,
 	toggleGroup,
 };