+++ packages/fields/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Enhancements
+
+- Fields: The `rich-text` and `title` fields now use the DataViews `richtext` control, removing the dependency on `@wordpress/rich-text-control`. [#78825](https://github.com/WordPress/gutenberg/pull/78825)
+
 ### Internal
 
 -   Migrate `Tooltip` consumers from `@wordpress/components` to the new compositional `Tooltip` in `@wordpress/ui` ([#78691](https://github.com/WordPress/gutenberg/pull/78691)).